FIxed `backoffError` mocks
